Update patch set 7
Patch Set 7: Code-Review-1

(3 comments)

I compared the add ports functionality to delete ports functionality and there appears to be a discrepancy. For this to work as a performance script I think we should be able to delete in the same way as we add using the script.

Patch-set: 7
Label: Code-Review=-1