Update patch set 8
Patch Set 8: Code-Review+2

ok, python good with marcus, robot good with me.  the verify job
has all the CSIT passing.  +2 and merge.

Cangji,

it would really be good to move away from the custom python REST 
library this project is using for CSIT and move to the standard
(for OpenDaylight CSIT) robot RequestsLibrary.  This allows a wider
audience to debug (aka help) and give Iotdm future benefit as 
our common libraries grow and improve.  If you can lobby to get
that done in Boron, it would be really appreciated.

Thanks for all the patience and iterations on this patch.  It does
go slow sometimes.

JamO

Patch-set: 8
Label: Code-Review=+2