Make pep8 more picky
Several guidelines were disabled and as a result I was
slipping into my old coding style, incompatible with the
Python sources here. Enabling them turned out to uncover I
was not the only one slipping. As a matter of fact all
checks are enabled now except E133 which is not compatible
with E123.

Change-Id: I754c740f49a213360df5b187897e776266e6653f
Signed-off-by: Jozef Behran <jbehran@cisco.com>