Clear known bug ID in SetupUtils.Teardown_Test_Show_Bugs_...
When a known bug ID is set to e.g. 1234 and the test steps
after that will catch the bug 1234, the bug ID 1234 is
reported (correctly). But then the rest of the test case is
skipped and the next one is started. This means that the
next test case will have the bug ID set to 1234 and if
it fails, it will also report it found bug 1234 while in
reality it was supposed to report a bug search link instead.
Fix the problem by clearing the bug ID after reporting the
bug.

Change-Id: I44d9c17bd9e22dc4792529dea0425a7f6d06940e
Signed-off-by: Jozef Behran <jbehran@cisco.com>