Update patch set 2
Patch Set 2: Code-Review-2

Thanks for this WIP. (I added -2 because it's a WIP)

I have opened an email discussion [0] to figure out the best way to
do docker things in CSIT.  I think with a dockerfile we will have to
build it still which might take time we don't want to spend on CSIT
jobs.  Maybe we can create images and upload them somewhere official 
and then just pull them down as needed.  We'll see.

The other thing I want to try to make happen is to move as much work 
up to the robot level as possible.  Our CSIT jobs are so much nicer to
debug, use and share when the work is done with robot keywords.  We'll see how far we can get with that.  I am *VERY* interested in 
getting this work upstream and running so I will help as much as
possible.

Also, we have some docker python library [1] already.  I have no idea
what it does, why it's there, or any other details.  just pointing it
out since it might help or conflict or be totally benign :)

[0] https://lists.opendaylight.org/pipermail/integration-dev/2016-March/006036.html
[1] https://github.com/opendaylight/integration-test/blob/master/csit/libraries/Docker.py

Patch-set: 2
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-2