Do not emit raw binary data after unconfirmed open
When the OPEN message generated by the player is rejected,
the original code emitted raw binary data in the resulting
exception. Change that so it emits the hexlified version of
the raw binary data. The fix was done by removing the
duplicated code that generates the two error messages (one
for the log file, the other for the exception), taking the
correct code only.

Change-Id: If1ac202ba0d3c8720c8c442da1d533143dcddb7c
Signed-off-by: Jozef Behran <jbehran@cisco.com>