Update patch set 2
Patch Set 2:

(1 comment)

One nitpick about the documentation.

This solution is "good enough" to get things rolling again on Releng. A better solution would be to have a list of candidate paths and query the connection until one of the paths fits. But I agree that this "better solution" shall be postponed until we will get to the point when another system with a third JAVA_HOME path comes along.

Patch-set: 2
Label: Code-Review=0