Add auth key for 0.0.0.0/0 in LISP perf tests
We recently discovered a security issue in lispflowmapping and an
explicitly configured authentication key is needed since the fix was
committed to allow southbound Map-Register messages to write into the
mapping database.  This patch adds such a key for the Map-Register tests
to produce usable results (notifies/s graph has value 0).

Change-Id: I14bac6c92f040abce71f40a24c531b9ca8e312b3
Signed-off-by: Lorand Jakab <lojakab@cisco.com>