Add more logging into Resolve_XXX_From_Template_Folder
This code is already logging the content of the template
itself. Here the logging of the resolved piece of data is
added as well. This helps especially when the "XXX" part is
"XYZ_Data" as the resulting data might get too large to be
logged in its entirety automatically on variable
assignments.

Change-Id: I3169df2d6ccd714a7df09aa3e6785b2528bd8d2b
Signed-off-by: Jozef Behran <jbehran@cisco.com>