Update patch set 23
Patch Set 23: Code-Review+1

(12 comments)

+1 for the suite doing what is should. I would like better variable names, but that is not a blocker.

But this suite is not compatible with Lithium (peer config data would need "flowspec" instead of "ipv4-flowspec").
Testplan in this Change assumes this suite would be run in a separate job, but I would like to make it part of "userfeatures" job. Either by making peer configuration try Lithium way if the default fails, or by leveraging Change 35774.
But I am not opposed to merging the bgpcep-flowspec testplan and making changes later.

Patch-set: 23
Label: Code-Review=+1