Update patch set 2
Patch Set 2:

> Hey so did we get some feedback regarding the docker containers and
 > where we could load them?
 > 
 > Also I've seen that the issue can be replicated only using mininet,
 > so maybe no need to spawn docker containers...
 > 
 > I might finally have some time to work on this next week :)

mininet will be the fastest route.  you could modify/enhance the 
current scale jobs [0].  I assume all you really need to worry 
about is getting ovs 2.4 right?  If so, we can change the mininet
VM that the scale job uses.  There is one problem though, which is
why we kept mininet using 2.0.  2.0 is predictable with how it 
names and uses it's ports.  So in a linear topo, each switch has three ports (1 to controller and 1 each to the switch on either side).  I'm not sure the exact number, but port 0 *always* connects to controller, s2p1 always goes to s1p2, etc.  Our scale test makes
this assumption.  That assumption is not right with ovs2.4.

So, we'd need to modify the test to account for this, or create a new
job (my preference) that uses mininet with 2.4. Then enhance the same
robot suite to either (1) just skip the link checking (this is what 
breaks with the port unpredictability) or (2) enhance the link checking so that it works.

make sense?  I can hopefully give as much pointers as you need to get it working.


-or-

we can do docker.  there is a docker VM in the LF releng cloud.  I've
used it successfully in the sandbox to bring up 3 ODL containers.  it
was just a POC, but it worked.  There is a tool installed called sshuttle that would even allow the containers to accessed from another
VM which would run your controller.

let me know what you think?

JamO

Patch-set: 2