Update patch set 12
Patch Set 12:

I have two patches [0,1] that should help make these jobs a little
more stable, but there is one unsolved problem, but I started an 
email [2] for that.

I started a sandbox job [3] that will use this patch and use the right
values for ODL_VERSION to hopefully give us the confidence we want to
merge this patch.  let's see how it goes.

[0] https://git.opendaylight.org/gerrit/#/c/36518/
[1] https://git.opendaylight.org/gerrit/#/c/36519/
[2] https://lists.opendaylight.org/pipermail/integration-dev/2016-March/006168.html
[3] https://jenkins.opendaylight.org/sandbox/job/ovsdb-openstack-tempest-mitaka-beryllium/1/

Patch-set: 12