Add test for RestPerfClient timeouts
When RestPerfClient times out, it means a poor Netconf
Connector performance bug was caught. In the MDSAL Northbound
case it is reported as bug 5581, no such bugs were seen in
the other two cases so none are reported. However after the
RestPerfClient bug that causes it to hang in this situation
(see bug 5413) is fixed, the suite in itself tends to fail to
take a notice of such problems so add this specific check for
timed out threads that points the poor Netconf performance
bugs out.

Change-Id: Ib75707f517ee34f40b6c9961f90309c26870bafb
Signed-off-by: Jozef Behran <jbehran@cisco.com>