Update patch set 15
Patch Set 15:

Hi Jamo & Venkat,
the two tests which are ignored currently are not needed for lithium snapshot. The tests are added for verification of flows in Beryllium.
So if we run the patch using lithium snapshots they are ignored.
Please let us know in case of any queries.

Patch-set: 15