Update patch set 1
Patch Set 1: Code-Review+2

This fixes the falsely failing test.
Now that we see the data, the test could be improved to verify loc-rib contains all the expected data. For example use more specific URL and then use norm_json.py to normalize and compare json.
But that can be done in a separate Change.

Patch-set: 1
Label: Code-Review=+2