Export HTML status checking as a public keyword
Checking the HTML status for "valid success values" is quite
complicated with accessing a global variable holding a list
of acceptable values. Export this functionality for external
use without exposing this internal detail. Complements the
change that allows the callers of PutXmlTemplateViaRestconf
to request the status code to be returned.

Change-Id: Idceb864c4d8134093570c5f604acfc507418adbd
Signed-off-by: Jozef Behran <jbehran@cisco.com>