Update patch set 3
Patch Set 3: Code-Review+1

(2 comments)

The Python looks very nice, well-built. A few suggestions about places you could use existing libs inline. In general docstrings are nice, but your methods are so small and do-one-thing so well it's not as necessary as a lot of code we review. In general we recommend PEP8s prescribed naming convention, which would use lowercase/snake_case and avoid Cap Vars/CamelCase. However, you clearly have picked a style and been consistent, so imho this is okay. The commit message is outstanding.

Patch-set: 3