Add clustering performance test suite
This is a suite which deploys a device and puts some data
into it and then issues requests to update the data. The
difference is that the device is configured and the data
is added using one node and the data modification requests
are issued on another node of the cluster, forcing the
cluster to do clustering communication.

Change-Id: I902d2b9aa924de7e1f9fd891600387f769b91013
Signed-off-by: Jozef Behran <jbehran@cisco.com>