Update patch set 4
Patch Set 4:

> Hi Jamo,
 > 
 > I am unaware about the process of sandbox testing. So yes, I will
 > be glad, if you could please provide some guidance and useful links
 > for creating a sandbox example of this test code.
 > 
 > Thanks,
 > Ruchita.

Ruchita,

the learning curve for the sandbox is (well used to be) a little
steep.  just fyi, but lots of people are using it lately so I think
it must be getting easier to do.

anyway, start here [0] and let me/us know if you need help.

the summary of how it works.

- pull the releng repo
- push a test job to the sandbox
    * for you, you'll have no editing to do; just push your centinel-basic job to the sandbox.
- start the test suite, but use this test patch for PATCHREFSPEC in the build parameters
    * e.g. for patch set 4, it would be "refs/changes/40/36940/4"
- copy the link to the results back to this gerrit as proof that
    your test code is doing what you wanted.

The whole process, once you get the hang of it, is only a matter 
of tens of minutes to get results.

oh, you need to get access to the sandbox.  just an email to help desk,
but that info should be in the wiki I gave for [0]

thanks for the contributions :)

[0] https://nexus.opendaylight.org/content/sites/site/org.opendaylight.releng.builder/master/jenkins.html#jenkins_sandbox

Patch-set: 4
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0