Replaced NetconfViaRestconf.robot with TemplatedRequests.robot
NetconfViaRestconf.robot is not compatible with
ClusterManagement.robot and is obsoleted by
TemplatedRequests.robot. The incompatibility problem prevents
easy construction of test suites testing scenarios with node
outages so this outdated piece of code had to be replaced
with the newer (and compatible) piece of code. The change is
quite involved because TemplatedRequests.robot has quite
different API.

Change-Id: Iaa791eab06d1eaaf4d1904f248c57f67a30afc61
Signed-off-by: Jozef Behran <jbehran@cisco.com>