Fix spurious warnings after a peer is killed
When a peer is killed and is later restarted, then any
attempt to query anything from it using restconf will lead
to a bunch of warnings about a dead connection. After digging
in the RequestsLibrary source I realized that the proper cure
is to tell the RequestsLibrary.Create_Session to not make any
retries if the first connection attempt fails.

Change-Id: Ic0586a280d2d9bda675771182f66374c4c6cd982
Signed-off-by: Jozef Behran <jbehran@cisco.com>