Update patch set 19
Patch Set 19: Code-Review+1

Well, changed my mind about this "code duplication". I was trying to reuse this code for my test suite but I only discovered that I can't make it work in a reasonable amount of time so I am going to end up duplicating the code as well. So let's merge this as it is but with a TODO telling that the keywords need merging.

Patch-set: 19
Reviewer: Gerrit User 2877 <2877@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1