Update patch set 44
Patch Set 44:

> Hi Jamo,
> 
> There are basically 2 major reasons for having python libraries "Openstack.py" and "Vpnutils.py" instead of 
> using the robotrequests library.
> 
> 1) Scalability and Performance: If we use the robotrequests library, we might have to maintain jsons files for
> multiple entity creation. Currently we are building the json dynamically in the python script.  Going forward 
> if at all there is a change in json structure, this wrapper helps to adapt the changes as we pass the required 
> parameters in the python function. And this will improve the performance during bulk creation

I'm missing the scalability or performance test in this patch.  We have allowed duplicated
work in python for this kind of stuff in the past.  I guess this would be an example library [0]
used by a .robot suite [1]

[0] https://github.com/opendaylight/integration-test/blob/master/csit/libraries/ScaleClient.py
[1] https://github.com/opendaylight/integration-test/blob/master/csit/suites/openflowplugin/Performance_Flow_Config/090_BulkMaticConfigDsSetup_100k.robot


However, this is kind of a one-off job for the sole purpose of scaling openflow flows.


If your patch is for something like that, then I think we can approach it differently,
otherwise I still argue to move all REST work to use robot libraries.  It allows us to
share and collaborate.  If this patch goes upstream, all initial authors go away and
then it breaks.  Someone will have to learn, debug and maintain this.  This is one way
to make that easier for those future maintainers.  We have some openstack related libraries
already, that may or may not do some of the things you need.  But, you could update
those to do whatever you need, and that in turn will be easily available for others
to take advantage of.


As for dealing with JSON structure we have two methods folks are using upstream
already.  maybe you would find them useful?

one is to read a .json file and then "Replace String" with whatever values
you need at that point in your suite.


> 2) Multipurpose library:  Enhancement for openstack script has the start and stop VM functions with tapport and 
> running the more number of VMs in multi thread environment.

I missed this library.  which file is it?

> Except the above concern, all other review comments are addressed.
> Could you please clarify this is the only comment/concern that you have? or any other review comment blocks 
> from merging?

Even with the above conversation, I have more comments to put in line.



I understand this is an already mature test you are now trying to get upstream,
so it's much larger than and more painful to get merged.  I'm hoping others will
comment as well, so we can help get things the way everyone sees best.

Thanks,
JamO

> 
> 
> Thanks,
> Mahendar.

Patch-set: 44