Adding ElasticsearchAppenderClass & longevitytests
This adds ElasticsearchAppender Class which can query for harvested
MBeans in an Elasticsearch Node to which karaf-decanter dumps the data.
The current MBeans that can be currently queried for are:
    java.lang:type=GarbageCollector
    java.lang:type=Threading
    java.lang:type=ClassLoading
    java.lang:type=Memory

This also adds a method plot_points which draws graph using mathplotlib
for the above metrices
    Usage example is defined in the file.

This also adds long duration robot tests for the above class in the
folder longevity. A sample run would look something similar to

robot -v PORT:9200 -v IP:10.2.4.204 -v DURATION:20 -v STEP:1 010_Check_JVM_Resource.robot
where elasticsearch node is running on 10.2.4.204:9200 and each of the
tests is to be run for a duration of 20 seconds.

Change-Id: I4acb3cef32213e91cbaadb40c0c21deef8f4c060
Signed-off-by: Kumar Rishabh <shailrishabh@gmail.com>