Update patch set 3
Patch Set 3:

> > can we get an exception from pep8 style check?
 > 
 > My initial response is "no".
 > 
 > I believe that for generated code in Java projects, it is either
 > covered by Checkstyle or never present in git (only in source
 > jars).
 > So I suggest either to generate the binding in during the job run
 > (possibly using a scriptplan), or to clean up the generated code
 > manually (so we can merge it).

Cleaning up manually is not an option, since we don't want to do that every time this is re-generated (even if that doesn't happen much). We could use automatic formatters, but I personally don't usually like the results.

So let's go to the run-time generation option. What is a scriptplan?

Patch-set: 3
Label: Verified=0