Update patch set 33
Patch Set 33:

> (3 comments)
 > 
 > Sandbox shows two issues. Firstly, Check Count Integrity
 > exclusively uses member-1 as source of its truth, but member-1 got
 > hit by Isolate Switchs Candidate [0]. Secondly, Isolate Switchs
 > Candidate passed (maybe a Bug in ODL, but maybe just a luck of
 > Robot detecting an intermediate state) while the following Rejoin
 > Switchs Candidate failed instead (probably a Bug in ODL).
 > The first issue should be fixed in the suite by selecting
 > appropriate member for getting the list of candidates (or by
 > querying all members). The second issue should be examined (if it
 > is still visible after fixing the first issue).
 > 
 > [0] https://jenkins.opendaylight.org/sandbox/job/openflowplugin-csit-3node-clustering-only-beryllium/4/robot/report/log.html#s1-s4-t2-k1-k5

member-1 is not used as source of truth exlusively, variable ${active_session} is set for it. 
Kw  Isolate Switchs Old Owner - line 95 sets it before node is isolated, lines 99 and 103 after the isolation(setting a session to a new masster). After that there is no need to change it, because we work only with followers. So failure in the test is probably a bug.

Patch-set: 33