More readable log message for ExecuteCommandOnRemoteSystem
The original logging command produced a message like
"Attempting to execute sudo /sbin/iptables -I OUTPUT -p all
--source 10.30.32.162 --destination 10.30.32.204 -j DROP on
10.30.32.162 by jenkins with any and >" which is pretty
confusing as it is not clear what part is the command to be
execute, what part is the remote system to be executed and
so on.

The new format is something like 'Attempting to execute
command "sudo /sbin/iptables -I OUTPUT -p all --source
10.30.32.162 --destination 10.30.32.204 -j DROP" on remote
system "10.30.32.162" by user "jenkins" with keyfile "any"
and prompt ">"' which is far more readable as now the
message described each of the items it contains and also
uses quotes to make it explicit where the items start and
end in the message.

Change-Id: I886059bd7bae6ebde866977a49b4bf6eb63354d1
Signed-off-by: Jozef Behran <jbehran@cisco.com>