BGP 10 peers: Lower scale in order to fit in memory.
With Route Reflector client peers, all data sent to them
is is AdjRibsOut, which contains copies for each peer.
That means there was 9M more prefixes stored in datastore,
which is too much for the heap to handle.

This sets scale to 100k prefixes, which should fit in memory safely;
unless there is large amount of transient data during processing.

Change-Id: I11c1a3d4fa966aee56145949520918c4307d8335
Signed-off-by: Vratko Polak <vrpolak@cisco.com>