Fix MDSAL suite failing if run on a VM with multiple roles
When the VM in which the MDSAL suite is running has multiple
roles (it is ODL host as well as the tools system to be
exact), the suite fails because the command that tries to
determine the PID of the ODL Netconf connection finds
multiple "sshpass" processes and causes all test cases that
need a PID to fail, including the suite teardown. Fix this
by determining the TTY ID of the SSH connection that will
become the Netconf connection and filtering the processes
by this TTY ID as well.

Change-Id: Ia61ef9e9a64b08a5b3c98c642e6888cd2a343b0c
Signed-off-by: Jozef Behran <jbehran@cisco.com>