Fix broken JAVA_HOME detection
The SSHLibrary.ExecuteCommand keyword runs the command in a
shell that fails to configure its environment correctly (most
likely by failing to run any startup files located in the
account to which the connection is logged in), hiding the
JAVA_HOME value and crashing the ComposeBaseJavaCommand.

Add SSHUtils.Invoke keyword which fixes this problem and
modify the broken code to use that instead. The original
SSHLibrary.ExecuteCommand keyword shall be avoided and
SSHUtils.Invoke shall be used to avoid additional nasty
surprises like this one.

Change-Id: I45418c7406d5933b23f8f712241d5cfa44470d14
Signed-off-by: Jozef Behran <jbehran@cisco.com>