Fix broken JAVA_HOME detection
The SSHLibrary.ExecuteCommand keyword runs the command in a
shell that fails to run any startup files located in the
account to which the connection is logged in, hiding the
JAVA_HOME value and crashing the ComposeBaseJavaCommand. Add
code to trick the connection into revealing the value of
JAVA_HOME by issuing the command querying for the JAVA_HOME
using the SSHLibrary.Write keyword, asking the command to put
the value into a file and then using SSHLibrary.ExecuteCommand
to dump the file to a Robot variable.

Change-Id: I45418c7406d5933b23f8f712241d5cfa44470d14
Signed-off-by: Jozef Behran <jbehran@cisco.com>