Fix spurious failures on small BGP batches
The suites manypeers_changecount.robot and
manypeers_prefixcount.robot are failing if too small amount
of prefixes is being specified because the timeout
calculations are not taking static overhead into account.
This static overhead is too high for small amounts of
prefixes, leading to spurious failures due to a timeout.

Change-Id: Ib46f4dc099bb1c55d95422f6ecee4ee8b8e45bf3
Signed-off-by: Jozef Behran <jbehran@cisco.com>