Add replacement for SSHLibrary.PutDirectory to SSHKeywords
Using SSHLibrary.PutDirectory to deploy stuff onto the
virtual machines has undefined behavior when the destination
directory already exists. Additionally, the name of the
keyword ("SSHLibrary.PutDirectory") suggests that in this
situation the content of the destination directory will be
replaced with the data in the source directory (think about
HTTP PUT requests if you are unsure why that is so) but
that is not the case.

This adds two replacement keywords with more readable names
and defined (and documented) behavior. The first one, named
SSHKeywords.DeployDirectory is to be used for data directory
deployment and it will replace any existing directory with
the content of the source directory. The second one, named
SSHKeywords.UpdateDirectory will merge the content from the
source directory into the data in the destination directory
if the destination directory already exists.

Change-Id: Ia87e8e133f5e158aa3150d1a09597669b6f05ec8
Signed-off-by: Jozef Behran <jbehran@cisco.com>