Update patch set 3
Patch Set 3:

> Patch Set 3: -Code-Review

This won't work;  the bug involves the fact that the token will be granted (i.e., you will get a token) but that the token wasn't stored appropriately in ehcache.  Thus, the request to get the token will succeed;  you just won't be able to use that token.

Patch-set: 3