Update patch set 8
Patch Set 8:

Some test cases fail because of bug 5992, but this change brings us closer to causes of alike bugs. Here notice, that testcase "On GBPSFC1 Send HTTP req h35_2-h36_2 to GBPSFC2" passes because it behaves as it should. Currently it's failing too and it's harder to determine which of the nodes in transaction chain is breaking the chain.

Patch-set: 8