Update patch set 2
Patch Set 2: Code-Review+1

I think this needs a manual rebase, but no comments from me on
the code review.

do you have the jenkins job part of this ready too?  You can push
it to the sandbox and run with this patch there to get an idea of
what it looks like.

Patch-set: 2
Label: Code-Review=+1