Update patch set 1
Patch Set 1: Code-Review+2

merging this, but the other option (which I personally prefer) is to let it keep failing to keep visibility to the outside that we have known bugs that need to be fixed.  if you ever want to do that, just add a line like this [0] to the end of the relevant test cases.  that
makes it easier for us to know it's a well know bug/failure.

[Teardown]    Report_Failure_Due_To_Bug    ####

where #### is the actual bugzill number

Patch-set: 1
Label: Code-Review=+2