Update patch set 1
Patch Set 1: Code-Review+1

This is weird.  This patch is removing the extra username/password args to the Run Keyword On Remote System, which works in the normal CSIT job, as I tested in sandbox [0].  But this change was not needed to fix releng CSIT, as that was working before this patch [1].  I'm hoping it fixes the verify csit job [2] though, which Victor is seeing fail.  I don't know why the verify job would fail with this, but not normal releng csit.  I don't think this patch would hurt, as almost all instances of this keyword in our repo do not give the username and password explicitly.

[0]  https://jenkins.opendaylight.org/sandbox/job/usc-csit-1node-tcp-only-boron/
[1]  https://jenkins.opendaylight.org/releng/view/usc/job/usc-csit-1node-tcp-only-boron/
[2]  https://jenkins.opendaylight.org/releng/job/usc-csit-verify-1node-tcp/23/

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1