Update patch set 1
Patch Set 1:

> Looks like a pretty infrastructure file. Should've this be a
 > parameter with default value (INFO) where the value could be set by
 > an environment variable? This way a specific Jenkins run could set
 > value to TRACE at any time.

we can pass "-L TRACE" on the pybot command line (which is available in most jobs
as a jenkins parameter with the TESTOPTIONS), but that would affect all keywords.

we could also modify this karaf keyword to have an argument ${loglevel}=DEBUG and
then override it when needed, but that would require us to track down all the places
we want to do this and override it with TRACE.

we could also try the environment variable idea, where loglevel=%ROBOTLOGLEVEL
instead, and then go tweak the jenkins scripts that run pybot to also add that
environment variable.

this was just a quick update to try to help figure out the EOFError problem we
sometimes run in to, and was thinking we'd revert it once we figured it out, so
didn't want to spend a lot of time being fancy :)

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0