Update patch set 8
Patch Set 8:

This is all happening in the robot VM I assume. Since this is a Jenkins slave, it is normally recommended to place files in the ${WORKSPACE} instead of /tmp. This way things get automatically cleared from one run to another + you can debug better in the sandbox by looking at the WORKSPACE. BTW, have you tried this in sandbox?

Patch-set: 8
Label: Verified=0