Update patch set 6
Patch Set 6:

(6 comments)

-1 for "followers vs candidates", and perhaps also for not skipping the default "all members" list.
Other comments are not that relevant.
Will look more closely at the suites when the two issues are addressed.

Sorry but I would prefer not to call device candidate = follower for the reason I explained in the patch. I think I am going to bring this to the mail list to see what other committers think.

Patch-set: 6