Update patch set 6
Patch Set 6:

OK, after talking to cluster devs, it seems like candidates are really instances that register for ownership and the owner is elected between the candidates. We can keep this definition in this KW to match the API behavior and i will create another KW that gives the candidates without the owner. Still I do not think it is good idea to call this follower because of the overlap with leader/follower definition but lets bring this last question to the list to see what people think.

Patch-set: 6
Label: Verified=0