Update patch set 2
Patch Set 2: Code-Review+2

first of all, than you so much for the small and easily review-able patch :)

secondly, normally we like to see these patches with a sandbox run to prove that it's ok, but I know that is a bigger learning curve and I think we can deal with this patch being merged now if there are any 
troubles in releng.

maybe you are already ok with the sandbox, but if not, I recommend
that you get familiar with it if you are planning to add more 
system test, which I hope.

good work.

Patch-set: 2
Label: Code-Review=+2