Update patch set 1
Patch Set 1:

(1 comment)

this patch is good and makes things more flexible.  the goal of the original patch was only to quickly try to figure out why we are getting that EOFError sometimes in the new cloud.  With this patch, we'll need to find all the consumers of this keyword and have them use the TRACE argument.

Patch-set: 1
Label: Code-Review=0