Update patch set 9
Patch Set 9:

(6 comments)

> This is all happening in the robot VM I assume. Since this is a
 > Jenkins slave, it is normally recommended to place files in the
 > ${WORKSPACE} instead of /tmp. This way things get automatically
 > cleared from one run to another + you can debug better in the
 > sandbox by looking at the WORKSPACE. BTW, have you tried this in
 > sandbox?

Hi Luis, 

I tried to run it in Jenkins sandbox but it is failing as this script requires pyangbind module to be installed in the virtualenv, I asked it on the integration-dev list about the installation of module and its dependencies, It seems that there is some issue with the current robot image. I have spent hours in figuring out a way of installing module by adding commands in executable shell parameters in config tab of UI but It is not allowing me to install system level dependencies on lxml libxml2-dev and libxslt1-dev.

Link to thread: https://lists.opendaylight.org/pipermail/integration-dev/2016-July/007454.html 
Link to Jenkins job I tried: https://jenkins.opendaylight.org/sandbox/job/lispflowmapping-csit-1node-msmr-only-boron/

I have made the changes as you mentioned, Please review it. Also pyangbind should be installed in the production VM images for the tests to use it.

Patch-set: 9
Label: Verified=0