Change controller PID regex
Controller PID regex was: java.*distribution.*karaf

This regex will work only if 'distribution' is part of path name,
but if you installed opendaylight in '/opt/opendaylight', for
example, it will fail.

The only strings that can be always found in the PID regex for sure,
are 'java' and 'karaf'.

Change-Id: Ie923297745e863d74a900c2145e4122d834ea2c6
Signed-off-by: Arie Bregman <abregman@redhat.com>