Update patch set 1
Patch Set 1: Code-Review+1

Jamo/Luis is there a way to run integration/test changes in sandbox, so we can have a preview of these results?

In any case according to our downstream on old netvirt this commit will add a bunch of passing tests, a few skips and 6 failing tests.

Of the failing tests, 2 check security group functionality (which is actually missing and is a bug), 2 check port/route admin state (I believe this to also be missing, not sure if there is a bug for it), and 2 check IPv6 which is currently unsupported.

We might want to exclude tempest.scenario.test_network_v6 if there's no interest in seeing these fail.

Patch-set: 1
Reviewer: Gerrit User 670 <670@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1