Update patch set 1
Patch Set 1:

> Jamo/Luis is there a way to run integration/test changes in
 > sandbox, so we can have a preview of these results?
 
Yeah, we definitely need this.  In my initial work to get just tempest.api.network
running I was not even able to get this test_minimum_basic to work.  I did not
focus on it though as just getting api.network working was my first goal.

push one of the openstack-$openstackversion-openstack-$odlversion jobs
to sandbox and then run with this patchrefspec to see what it looks like.
paste the link to that job here and it will be easy for us to check it out
to help debug.

 > In any case according to our downstream on old netvirt this commit
 > will add a bunch of passing tests, a few skips and 6 failing tests.
 > 
 > Of the failing tests, 2 check security group functionality (which
 > is actually missing and is a bug), 2 check port/route admin state
 > (I believe this to also be missing, not sure if there is a bug for
 > it), and 2 check IPv6 which is currently unsupported.
 > 
 > We might want to exclude tempest.scenario.test_network_v6 if
 > there's no interest in seeing these fail.

if v6 is not a feature we are going to give in ODL then skipping is fine,
but having failures is fine with me if we have bugs to track them.

Thanks for looking at this guys.

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0