Update patch set 9
Patch Set 9:

> > (1 comment)
 > 
 > do we not have to restart the controller when we make log level
 > modifications in the logging config file?


Jamo,
I think It is not necessary, with the change of this patch we are getting Loglevel as Trace as we expected.
For your reference below are the karaf.log file with TRACE enabled for VTN.
1) https://logs.opendaylight.org/sandbox/jenkins091/vtn-csit-1node-manager-only-boron/7/archives/karaf.log.gz
2) https://logs.opendaylight.org/sandbox/jenkins091/vtn-csit-1node-manager-only-beryllium/6/archives/karaf.log.gz

So, please review and approve the patch. Since we are getting failure again and again due to karaf.login issue in below.
https://logs.opendaylight.org/releng/jenkins092/vtn-csit-1node-manager-all-boron/203/archives/log.html.gz#s1-s1

Patch-set: 9