Update patch set 1
Patch Set 1:

> (1 comment)

Yes Jamo, It is returning Network ID. The same we can get from the get Net ID keyword. Verbose is not required as the purpose of the keyword "create Network" is to create network only and no need to have any other functionality in that as we have separate keyword for doing that.
  
Thank you for looking into this.

Patch-set: 1
Reviewer: Gerrit User 6148 <6148@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0