Update patch set 1
Patch Set 1:

(1 comment)

> > (1 comment)
 > 
 > Yes Jamo, It is returning Network ID. The same we can get from the
 > get Net ID keyword. Verbose is not required as the purpose of the
 > keyword "create Network" is to create network only and no need to
 > have any other functionality in that as we have separate keyword
 > for doing that.
 > 
 > Thank you for looking into this.

Ok.  I added Venkat and Priya to this review as I think one of them is how we got the verbose option in that keyword.  If it's ok with them to 
remove it, then we can merge.  I was just worried they might need
it for some debugging purposes.

Thanks.

Patch-set: 1