Update patch set 11
Patch Set 11:

(2 comments)

> (1 comment)
 > 
 > Is there a Sandbox run for this (for next week)?
 > I am not clear on how the detection of .yang files works (Beryllium
 > and Boron).

I am trying to run it in the sandbox but it is failing [1]

These tests are now using JsonGenerator.py library [2] and JsonGeneratory.py file is using the python classes generated in odl-lispflowmapping-yang-files generated at [3].

In generation of these python classes I need the corresponding yang files which I am pulling from master branch. Is it ok? or do I need to pull it from different branches like Be, Boron according to the environment. If yes, Can you please tell me the elegant way of doing it.  

[1] https://jenkins.opendaylight.org/sandbox/job/lispflowmapping-csit-1node-msmr-only-boron/5/console
[2] https://git.opendaylight.org/gerrit/#/c/38700/
[3]https://git.opendaylight.org/gerrit/#/c/43941/5/csit/suites/lispflowmapping/scripts/generate_bindings.sh

Patch-set: 11
Label: Verified=0