Update patch set 5
Patch Set 5:

scriptplan is normally excuted before ODL SW is downloaded to controller VM. For this change I would recommend configplan which is executed after controller is deployed in VM and before it is starts for the first time. The config xml files are always packed in the distribution so it is easy to edit them there if you want to skip 1 reboot.

Patch-set: 5
Label: Verified=0